Move props.types upper 1 level
