+++ packages/chakra-ui-core/src/CFlex/utils/flex.props.js
@@ -1,4 +1,4 @@
-import { SNA } from '../../config/props/props.types'
+import { SNA } from '../../config/props.types'
 
 const flexProps = {
   as: {
