+++ packages/chakra-ui-core/src/CMenu/CMenu.stories.js
@@ -1,7 +1,6 @@
 import { storiesOf } from '@storybook/vue'
 import { CMenu, CFade, CMenuGroup, CMenuButton, CMenuList, CMenuOptionGroup, CImage, CMenuItemOption, CMenuItem, CMenuDivider, CIcon, defaultTheme } from '..'
 
-console.log({ defaultTheme })
 
 storiesOf('UI | Menu', module)
   .add('With internal state', () => ({
