+++ packages/chakra-ui-core/src/config/index.js
@@ -1,0 +1,1 @@
+// to keep rollup work
