+++ packages/chakra-ui-core/src/config/props.types.js
